requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
uuid==1.30
numpy==1.23.5
scipy==1.10.1
matplotlib==3.8.0
PyQt5==5.15.10
Pillow==9.5.0
opencv-python==4.7.0.72
tensorflow==2.14.0
networkx==3.2.1
imageio==2.33.1
potracer==0.0.4